requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
Django==2.2.6
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
Markdown==3.1.1
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
PySocks==1.7.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tqdm==4.36.1
traitlets==4.3.3
tweepy==3.8.0
urllib3==1.25.6
wcwidth==0.1.7
